Combine project date fields
